Move getBlockRoot methods to BeaconStateAccessors
