+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/altair/SpecLogicAltair.java
@@ -97,7 +97,6 @@ public class SpecLogicAltair extends AbstractSpecLogic {
         new BeaconStateUtil(
             config,
             schemaDefinitions,
-            validatorsUtil,
             committeeUtil,
             predicates,
             miscHelpers,
@@ -136,7 +135,8 @@ public class SpecLogicAltair extends AbstractSpecLogic {
             attestationValidator,
             operationValidator);
     final ForkChoiceUtil forkChoiceUtil =
-        new ForkChoiceUtil(config, beaconStateUtil, attestationUtil, blockProcessor, miscHelpers);
+        new ForkChoiceUtil(
+            config, beaconStateAccessors, attestationUtil, blockProcessor, miscHelpers);
     final BlockProposalUtil blockProposalUtil =
         new BlockProposalUtil(schemaDefinitions, blockProcessor);
     final SyncCommitteeUtil syncCommitteeUtil =
