+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/phase0/SpecLogicPhase0.java
@@ -91,7 +91,6 @@ public class SpecLogicPhase0 extends AbstractSpecLogic {
         new BeaconStateUtil(
             config,
             schemaDefinitions,
-            validatorsUtil,
             committeeUtil,
             predicates,
             miscHelpers,
@@ -125,7 +124,8 @@ public class SpecLogicPhase0 extends AbstractSpecLogic {
             attestationValidator,
             operationValidator);
     final ForkChoiceUtil forkChoiceUtil =
-        new ForkChoiceUtil(config, beaconStateUtil, attestationUtil, blockProcessor, miscHelpers);
+        new ForkChoiceUtil(
+            config, beaconStateAccessors, attestationUtil, blockProcessor, miscHelpers);
     final BlockProposalUtil blockProposalUtil =
         new BlockProposalUtil(schemaDefinitions, blockProcessor);
