+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/phase0/statetransition/epoch/ValidatorStatusFactoryPhase0.java
@@ -75,7 +75,7 @@ public class ValidatorStatusFactoryPhase0 extends AbstractValidatorStatusFactory
                   updates.previousEpochTargetAttester = true;
 
                   updates.previousEpochHeadAttester =
-                      beaconStateUtil
+                      beaconStateAccessors
                           .getBlockRootAtSlot(state, data.getSlot())
                           .equals(data.getBeacon_block_root());
                 }
